# backend/contracts/openapi-v1.yaml
openapi: 3.1.0
...
info:
  title: 北京 AI 旅行规划助手 API
  version: 1.0.0
  description: |
    正式对话式 MVP 的唯一 HTTP 契约。所有 ID 都是不透明字符串；客户端不得解析 ID。
    日期时间统一使用 UTC ISO 8601。旧版 /api/v1/trips 暂时保留，但不属于本契约。
servers:
  - url: /
    description: 通过正式前端同源代理访问
...
paths:
...
  /api/v1/runs/{run_id}:
    get:
      tags: [Runs]
      operationId: getRun
      summary: 查询首次生成或修改任务的真实状态
      description: Run 已接受后的上游失败通过 status=failed 和 error 表达，本接口仍返回 200。
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/RunId'
      responses:
        '200':
          description: Run 当前状态
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunResponse'
        '401': {$ref: '#/components/responses/Unauthorized'}
        '404': {$ref: '#/components/responses/NotFound'}
        '500': {$ref: '#/components/responses/InternalError'}
  /api/v1/plans/{plan_id}:
    get:
      tags: [Plans]
      operationId: getPlan
      summary: 获取指定的不可变行程版本
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/PlanId'
      responses:
        '200':
          description: 指定 PlanVersion
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlanVersionResponse'
        '401': {$ref: '#/components/responses/Unauthorized'}
        '404': {$ref: '#/components/responses/NotFound'}
        '500': {$ref: '#/components/responses/InternalError'}
  /api/v1/photo-spots/featured:
    get:
      tags: [PhotoSpots]
      operationId: listFeaturedPhotoSpots
      summary: 获取首页北京出片灵感
      security: []
      parameters:
        - name: city
          in: query
          required: false
          schema:
            type: string
            const: 北京
            default: 北京
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 10
            default: 5
      responses:
        '200':
          description: 只包含已准入、来源完整的精选机位
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeaturedPhotoSpotListResponse'
        '422': {$ref: '#/components/responses/ValidationError'}
        '500': {$ref: '#/components/responses/InternalError'}
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: trip_session
  parameters:
...
    RunId:
      name: run_id
      in: path
      required: true
      schema: {$ref: '#/components/schemas/OpaqueId'}
    PlanId:
      name: plan_id
      in: path
      required: true
      schema: {$ref: '#/components/schemas/OpaqueId'}
...
  responses:
...
    Unauthorized:
      description: 缺少或无法验证匿名会话
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ApiErrorResponse'}
    NotFound:
      description: 资源不存在或不属于当前匿名用户
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ApiErrorResponse'}
...
    ValidationError:
      description: 请求可解析但字段或业务语义无效
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ApiErrorResponse'}
...
    InternalError:
      description: 未预期错误，不返回内部堆栈
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ApiErrorResponse'}
  schemas:
    OpaqueId:
      type: string
      minLength: 1
      maxLength: 128
      description: 不透明字符串 ID；消费者不得依赖前缀或把它转为数字。
    NullableOpaqueId:
      anyOf:
        - $ref: '#/components/schemas/OpaqueId'
        - type: 'null'
...
    RunKind:
      type: string
      enum: [initial_plan, revision]
    RunStatus:
      type: string
      enum: [queued, running, succeeded, failed]
    RunStageStatus:
      type: string
      enum: [waiting, running, succeeded, failed, skipped]
    RunStage:
      type: object
      additionalProperties: false
      required: [key, label, status, started_at, completed_at]
      properties:
        key:
          type: string
          enum:
            - understanding_request
            - resolving_pois
            - planning_routes
            - recommending_lodging
            - retrieving_photo_spots
            - validating
        label: {type: string, minLength: 1}
        status: {$ref: '#/components/schemas/RunStageStatus'}
        started_at: {type: [string, 'null'], format: date-time}
        completed_at: {type: [string, 'null'], format: date-time}
    RunError:
      type: object
      additionalProperties: false
      required: [code, message, recoverable]
      properties:
        code: {type: string, minLength: 1}
        message: {type: string, minLength: 1}
        recoverable: {type: boolean}
    Run:
      type: object
      additionalProperties: false
      required:
        - id
        - conversation_id
        - kind
        - status
        - trigger_message_id
        - base_plan_id
        - result_plan_id
        - current_stage
        - stages
        - error
        - created_at
        - started_at
        - finished_at
      properties:
        id: {$ref: '#/components/schemas/OpaqueId'}
        conversation_id: {$ref: '#/components/schemas/OpaqueId'}
        kind: {$ref: '#/components/schemas/RunKind'}
        status: {$ref: '#/components/schemas/RunStatus'}
        trigger_message_id: {$ref: '#/components/schemas/OpaqueId'}
        base_plan_id: {$ref: '#/components/schemas/NullableOpaqueId'}
        result_plan_id: {$ref: '#/components/schemas/NullableOpaqueId'}
        current_stage:
          type: [string, 'null']
          enum:
            - understanding_request
            - resolving_pois
            - planning_routes
            - recommending_lodging
            - retrieving_photo_spots
            - validating
            - null
        stages:
          type: array
          maxItems: 6
          items: {$ref: '#/components/schemas/RunStage'}
        error:
          anyOf:
            - $ref: '#/components/schemas/RunError'
            - type: 'null'
        created_at: {type: string, format: date-time}
        started_at: {type: [string, 'null'], format: date-time}
        finished_at: {type: [string, 'null'], format: date-time}
      allOf:
        - if:
            properties: {status: {const: queued}}
          then:
            properties:
              started_at: {type: 'null'}
              finished_at: {type: 'null'}
              error: {type: 'null'}
        - if:
            properties: {status: {const: running}}
          then:
            properties:
              started_at: {type: string, format: date-time}
              finished_at: {type: 'null'}
              error: {type: 'null'}
        - if:
            properties: {status: {const: succeeded}}
          then:
            properties:
              result_plan_id: {$ref: '#/components/schemas/OpaqueId'}
              error: {type: 'null'}
              finished_at: {type: string, format: date-time}
        - if:
            properties: {status: {const: failed}}
          then:
            properties:
              error: {$ref: '#/components/schemas/RunError'}
              finished_at: {type: string, format: date-time}
    Coordinate:
      type: object
      additionalProperties: false
      required: [latitude, longitude, coordinate_system]
      properties:
        latitude: {type: number, minimum: -90, maximum: 90}
        longitude: {type: number, minimum: -180, maximum: 180}
        coordinate_system: {type: string, enum: [GCJ-02, WGS-84, BD-09]}
    POI:
      type: object
      additionalProperties: false
      required: [poi_id, canonical_name, map_source]
      properties:
        poi_id: {$ref: '#/components/schemas/OpaqueId'}
        canonical_name: {type: string, minLength: 1}
        map_source: {type: [string, 'null']}
    RouteSegment:
      type: object
      additionalProperties: false
      required:
        - origin_poi_id
        - destination_poi_id
        - recommended_mode
        - duration_min
        - distance_km
        - cost_cny
        - walk_distance_m
        - transfers
        - reason
      properties:
        origin_poi_id: {$ref: '#/components/schemas/OpaqueId'}
        destination_poi_id: {$ref: '#/components/schemas/OpaqueId'}
        recommended_mode: {type: string, enum: [步行, 公共交通, 打车, 驾车]}
        duration_min: {type: integer, minimum: 0}
        distance_km: {type: number, minimum: 0}
        cost_cny: {type: [number, 'null'], minimum: 0}
        walk_distance_m: {type: [integer, 'null'], minimum: 0}
        transfers: {type: [integer, 'null'], minimum: 0}
        reason: {type: string, minLength: 1}
    ReferencePhoto:
      type: object
      additionalProperties: false
      required: [image_id, storage_url, thumbnail_url, source_id]
      properties:
        image_id: {$ref: '#/components/schemas/OpaqueId'}
        storage_url: {type: string, format: uri}
        thumbnail_url: {type: [string, 'null'], format: uri}
        source_id: {$ref: '#/components/schemas/OpaqueId'}
    SourceRef:
      type: object
      additionalProperties: false
      required: [source_id, source_platform, source_author, source_url, evidence_ids]
      properties:
        source_id: {$ref: '#/components/schemas/OpaqueId'}
        source_platform: {type: string, minLength: 1}
        source_author: {type: [string, 'null']}
        source_url: {type: string, format: uri}
        evidence_ids:
          type: array
          items: {$ref: '#/components/schemas/OpaqueId'}
    BestTime:
      type: object
      additionalProperties: false
      required: [type, display_text, applicable_seasons, applicable_weather, source_ids, confidence]
      properties:
        type: {type: string, enum: [固定时段, 日出关联, 日落关联, 亮灯后, 季节性]}
        display_text: {type: string, minLength: 1}
        applicable_seasons:
          type: [array, 'null']
          items: {type: string}
        applicable_weather:
          type: [array, 'null']
          items: {type: string}
        source_ids:
          type: array
          items: {$ref: '#/components/schemas/OpaqueId'}
        confidence: {type: number, minimum: 0, maximum: 1}
    PhotoSpotHit:
      type: object
      additionalProperties: false
      required:
        - spot_id
        - poi_id
        - spot_name
        - coordinate
        - location_description
        - positioning_guide
        - location_precision
        - reference_photos
        - best_time
        - source_refs
        - ingestion_status
        - confidence
      properties:
        spot_id: {$ref: '#/components/schemas/OpaqueId'}
        poi_id: {$ref: '#/components/schemas/OpaqueId'}
        spot_name: {type: string, minLength: 1}
        coordinate: {$ref: '#/components/schemas/Coordinate'}
        location_description: {type: string, minLength: 1}
        positioning_guide: {type: [string, 'null']}
        location_precision: {type: string, enum: [exact_poi, named_sub_poi, approximate]}
        reference_photos:
          type: array
          items: {$ref: '#/components/schemas/ReferencePhoto'}
        best_time:
          anyOf:
            - $ref: '#/components/schemas/BestTime'
            - type: 'null'
        source_refs:
          type: array
          items: {$ref: '#/components/schemas/SourceRef'}
        ingestion_status: {type: string, const: auto_verified}
        confidence: {type: number, minimum: 0, maximum: 1}
    ItineraryItem:
      type: object
      additionalProperties: false
      required:
        - item_id
        - poi
        - start_time
        - end_time
        - stay_duration_min
        - booking_reminder
        - entry_tip
        - route_from_previous
        - photo_spots
      properties:
        item_id: {$ref: '#/components/schemas/OpaqueId'}
        poi: {$ref: '#/components/schemas/POI'}
        start_time: {type: [string, 'null'], pattern: '^([01][0-9]|2[0-3]):[0-5][0-9]$'}
        end_time: {type: [string, 'null'], pattern: '^([01][0-9]|2[0-3]):[0-5][0-9]$'}
        stay_duration_min: {type: integer, minimum: 0}
        booking_reminder: {type: [string, 'null']}
        entry_tip: {type: [string, 'null']}
        route_from_previous:
          anyOf:
            - $ref: '#/components/schemas/RouteSegment'
            - type: 'null'
        photo_spots:
          type: array
          items: {$ref: '#/components/schemas/PhotoSpotHit'}
    DayOrigin:
      type: object
      additionalProperties: false
      required: [type, poi_id, name]
      properties:
        type: {type: string, enum: [user_lodging, recommended_area]}
        poi_id: {$ref: '#/components/schemas/OpaqueId'}
        name: {type: string, minLength: 1}
    ItineraryDaySummary:
      type: object
      additionalProperties: false
      required: [estimated_duration_min, dominant_transport_modes, estimated_walk_distance_m]
      properties:
        estimated_duration_min: {type: integer, minimum: 0}
        dominant_transport_modes:
          type: array
          uniqueItems: true
          items: {type: string, enum: [步行, 公共交通, 打车, 驾车]}
        estimated_walk_distance_m: {type: [integer, 'null'], minimum: 0}
    ItineraryDay:
      type: object
      additionalProperties: false
      required:
        - day_index
        - date
        - theme
        - start_time
        - end_time
        - origin
        - outbound_route
        - items
        - return_route
        - summary
      properties:
        day_index: {type: integer, minimum: 1, maximum: 5}
        date: {type: [string, 'null'], format: date}
        theme: {type: [string, 'null']}
        start_time: {type: [string, 'null'], pattern: '^([01][0-9]|2[0-3]):[0-5][0-9]$'}
        end_time: {type: [string, 'null'], pattern: '^([01][0-9]|2[0-3]):[0-5][0-9]$'}
        origin: {$ref: '#/components/schemas/DayOrigin'}
        outbound_route:
          anyOf:
            - $ref: '#/components/schemas/RouteSegment'
            - type: 'null'
        items:
          type: array
          minItems: 1
          items: {$ref: '#/components/schemas/ItineraryItem'}
        return_route:
          anyOf:
            - $ref: '#/components/schemas/RouteSegment'
            - type: 'null'
        summary: {$ref: '#/components/schemas/ItineraryDaySummary'}
    LodgingRecommendation:
      type: object
      additionalProperties: false
      required: [area_id, name, level, representative_station, reason, covered_attractions, avg_transit_min]
      properties:
        area_id: {$ref: '#/components/schemas/OpaqueId'}
        name: {type: string, minLength: 1}
        level: {type: string, enum: [首选, 备选, 当前住宿评估]}
        representative_station: {type: [string, 'null']}
        reason: {type: string, minLength: 1}
        covered_attractions:
          type: array
          items: {type: string}
        avg_transit_min: {type: [integer, 'null'], minimum: 0}
    Limitation:
      type: object
      additionalProperties: false
      required: [code, message]
      properties:
        code: {type: string, minLength: 1}
        message: {type: string, minLength: 1}
    ValidationCheck:
      type: object
      additionalProperties: false
      required: [code, severity, message, day_index, item_id, spot_id]
      properties:
        code: {type: string, minLength: 1}
        severity: {type: string, enum: [fail, warning, spot_fail]}
        message: {type: string, minLength: 1}
        day_index: {type: [integer, 'null'], minimum: 1, maximum: 5}
        item_id: {$ref: '#/components/schemas/NullableOpaqueId'}
        spot_id: {$ref: '#/components/schemas/NullableOpaqueId'}
    ValidationResult:
      type: object
      additionalProperties: false
      required: [status, checks, checked_at]
      properties:
        status: {type: string, enum: [pass, fail]}
        checks:
          type: array
          items: {$ref: '#/components/schemas/ValidationCheck'}
        checked_at: {type: string, format: date-time}
    PlannerInfo:
      type: object
      additionalProperties: false
      required: [model, model_version, prompt_version]
      properties:
        model: {type: [string, 'null']}
        model_version: {type: [string, 'null']}
        prompt_version: {type: string, minLength: 1}
    ItineraryPlan:
      type: object
      additionalProperties: false
      required:
        - schema_version
        - plan_id
        - request_id
        - status
        - title
        - overview
        - request_summary
        - lodging_recommendations
        - days
        - limitations
        - validation
        - planner
        - generated_at
      properties:
        schema_version: {type: string, const: 1.2.0}
        plan_id: {$ref: '#/components/schemas/OpaqueId'}
        request_id: {$ref: '#/components/schemas/OpaqueId'}
        status: {type: string, const: validated}
        title: {type: string, minLength: 1}
        overview: {type: string}
        request_summary: {type: object, additionalProperties: true}
        lodging_recommendations:
          type: array
          minItems: 1
          items: {$ref: '#/components/schemas/LodgingRecommendation'}
        days:
          type: array
          minItems: 1
          maxItems: 3
          items: {$ref: '#/components/schemas/ItineraryDay'}
        limitations:
          type: array
          items: {$ref: '#/components/schemas/Limitation'}
        validation: {$ref: '#/components/schemas/ValidationResult'}
        planner: {$ref: '#/components/schemas/PlannerInfo'}
        generated_at: {type: string, format: date-time}
    PlanVersion:
      type: object
      additionalProperties: false
      required:
        - id
        - conversation_id
        - run_id
        - version
        - base_plan_id
        - status
        - changed_days
        - change_summary
        - plan
        - retrieval_run_ids
        - knowledge_index_version
        - created_at
      properties:
        id: {$ref: '#/components/schemas/OpaqueId'}
        conversation_id: {$ref: '#/components/schemas/OpaqueId'}
        run_id: {$ref: '#/components/schemas/OpaqueId'}
        version: {type: integer, minimum: 1}
        base_plan_id: {$ref: '#/components/schemas/NullableOpaqueId'}
        status: {type: string, const: validated}
        changed_days:
          type: array
          minItems: 1
          uniqueItems: true
          items: {type: integer, minimum: 1, maximum: 5}
        change_summary:
          type: array
          items: {type: string}
        plan: {$ref: '#/components/schemas/ItineraryPlan'}
        retrieval_run_ids:
          type: array
          items: {$ref: '#/components/schemas/OpaqueId'}
        knowledge_index_version: {type: string, minLength: 1}
        created_at: {type: string, format: date-time}
...
    RunResponse:
      type: object
      additionalProperties: false
      required: [data]
      properties:
        data: {$ref: '#/components/schemas/Run'}
    PlanVersionResponse:
      type: object
      additionalProperties: false
      required: [data]
      properties:
        data: {$ref: '#/components/schemas/PlanVersion'}
    FeaturedPhotoSpot:
      type: object
      additionalProperties: false
      required:
        - spot_id
        - poi_id
        - poi_name
        - spot_name
        - cover_image
        - location_description
        - best_time
        - source
      properties:
        spot_id: {$ref: '#/components/schemas/OpaqueId'}
        poi_id: {$ref: '#/components/schemas/OpaqueId'}
        poi_name: {type: string, minLength: 1}
        spot_name: {type: string, minLength: 1}
        cover_image:
          anyOf:
            - $ref: '#/components/schemas/ReferencePhoto'
            - type: 'null'
        location_description: {type: string, minLength: 1}
        best_time:
          anyOf:
            - $ref: '#/components/schemas/BestTime'
            - type: 'null'
        source:
          anyOf:
            - $ref: '#/components/schemas/SourceRef'
            - type: 'null'
    FeaturedPhotoSpotListResponse:
      type: object
      additionalProperties: false
      required: [data]
      properties:
        data:
          type: array
          maxItems: 10
          items: {$ref: '#/components/schemas/FeaturedPhotoSpot'}
...
    ErrorDetail:
      type: object
      additionalProperties: false
      required: [field, code, message]
      properties:
        field: {type: [string, 'null']}
        code: {type: string, minLength: 1}
        message: {type: string, minLength: 1}
    ApiError:
      type: object
      additionalProperties: false
      required: [code, message, details, request_id]
      properties:
        code: {type: string, minLength: 1}
        message: {type: string, minLength: 1}
        details:
          type: array
          items: {$ref: '#/components/schemas/ErrorDetail'}
        request_id: {$ref: '#/components/schemas/OpaqueId'}
    ApiErrorResponse:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error: {$ref: '#/components/schemas/ApiError'}